+++ index.php
@@ -10,14 +10,14 @@ get_header();
 if ( have_posts() ) {
   while ( have_posts() ) {
     the_post();
-    get_template_part( 'template-parts/listing/articles/grid' );
+    get_template_part( 'views/listing/articles/grid' );
   }
 
   the_posts_pagination();
 
 } else {
 
-  get_template_part( 'template-parts/listing/articles/empty' );
+  get_template_part( 'views/listing/articles/empty' );
 
 };