+++ search.php
@@ -27,17 +27,13 @@ if ( have_posts() ) { ?>
 if ( have_posts() ) {
   while ( have_posts() ) {
     the_post();
-    get_template_part( 'template-parts/listing/articles/grid' );
+    get_template_part( 'views/listing/articles/grid' );
   };
 
-  the_posts_pagination(
-    array(
-      'screen_reader_text' => ' ',
-    )
-  );
+  the_posts_pagination();
 
 } else {
-  get_template_part( 'template-parts/listing/articles/empty' );
+  get_template_part( 'views/listing/articles/empty' );
 
 };