+++ single.php
@@ -10,7 +10,7 @@ get_header();
 if ( have_posts() ) {
   while ( have_posts() ) {
     the_post();
-    get_template_part( 'template-parts/single/post' );
+    get_template_part( 'views/single/post' );
   }
 }