+++ skin/assets/scripts/fonts.js
@@ -1,9 +1,0 @@
-/* global WebFont */
-
-import WebFont from 'webfontloader';
-
-WebFont.load({
-  google: {
-    families: ['Open Sans:400:latin-ext'],
-  },
-});