+++ skin/assets/scripts/helpers/general.js
@@ -1,5 +1,0 @@
-export const generalHelpers = {
-  escapeString(string) {
-    return string.replace(/([;&,.+*~':"!^#$%@[\]()=>|])/g, '\\$1');
-  },
-};