+++ skin/assets/scripts/index.js
@@ -1,8 +1,2 @@
-/* global process __dirname */
 import '@babel/polyfill';
 import 'whatwg-fetch';
-
-import './fonts';
-
-import './form-fields';
-import './components';