+++ template-parts/tracking/before-body-end.php
@@ -1,8 +1,0 @@
-<?php
-/**
- * Tracking codes Before body end
- *
- * @package Inf_Theme\Template_Parts\Tracking
- */
-
-get_template_part( 'template-parts/tracking/parts/google-analytics' );