+++ views/tracking/before-body-end.php
@@ -1,0 +1,8 @@
+<?php
+/**
+ * Tracking codes Before body end
+ *
+ * @package Inf_Theme\Template_Parts\Tracking
+ */
+
+// Add Before body end hooks here.