+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-# Change Log for the Eightshift WordPress Boilerplate
 
+# Change Log for the Eightshift WordPress Boilerplate
 All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
@@ -7,6 +7,28 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 ## [Unreleased]
 
 _No documentation available about unreleased changes as of yet._
+
+## [4.0.0] - TBA
+### Added
+- Implementing Eightshift-libs
+- Updating docs
+- Adding type hinting
+- Updating npm packages
+- New file for shortcodes
+- New file for manifest
+- Complete webpack config refactoring
+
+
+### Changed
+- Changing folder structure to src
+- 
+
+### Removed
+- Removing a lot of stuff that is not part of boilerplate like utilities and some hooks and implementations
+- Removed general helper
+- Removed object helper
+- Removed Media implementations of svg upload
+- Unecesery Javascrip files
 
 ## [3.0.1] - 2019-04-23